syncplay_backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
channels==4.0.0
channels-redis==4.2.0
redis==5.0.1
django-cors-headers==4.3.1
daphne==4.0.0
asgiref==3.7.2
uvloop==0.19.0
websockets==12.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9 